// src/types/enums/text-style.enum.ts
export enum TextStyle {
	HEADING1 = 'HEADING1',
	HEADING2 = 'HEADING2',
	HEADING3 = 'HEADING3',
	HEADING4 = 'HEADING4',
	HEADING5 = 'HEADING5',
	HEADING6 = 'HEADING6',
	SUBTITLE1 = 'SUBTITLE1',
	SUBTITLE2 = 'SUBTITLE2',
	BODY1 = 'BODY1',
	BODY2 = 'BODY2',
	CAPTION = 'CAPTION',
	OVERLINE = 'OVERLINE',
}